+++ examples/nuxt/server/middleware/federation.ts
@@ -1,0 +1,7 @@
+import { fedifyMiddleware } from "@fedify/nuxt";
+import federation from "../federation";
+
+export default fedifyMiddleware(
+  federation,
+  (event, request) => undefined,
+);